refactor: validate 위치 변경
